# api/docs/swagger.yaml
basePath: /api
definitions:
...
  docs.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  handlers.CreateFolderRequest:
    properties:
      name:
        type: string
      path:
        type: string
    type: object
...
  handlers.FolderStats:
    properties:
      fileCount:
        type: integer
      folderCount:
        type: integer
      path:
        type: string
      totalSize:
        type: integer
    type: object
...
host: localhost:3080
info:
  contact:
    email: support@simplecloudvault.io
    name: API Support
  description: |-
    Enterprise-grade cloud file sharing and management API.
    SimpleCloudVault provides comprehensive file management, sharing, and collaboration features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: SimpleCloudVault API
  version: "1.0"
paths:
...
  /folders:
    post:
      consumes:
      - application/json
      description: Create a new folder at the specified path
      parameters:
      - description: Folder creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Folder created successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Folder already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create folder
      tags:
      - Files
  /folders/{path}:
    delete:
      consumes:
      - application/json
      description: Delete a folder by path. Use force=true to delete non-empty folders.
      parameters:
      - description: Folder path
        in: path
        name: path
        required: true
        type: string
      - description: Force delete non-empty folder
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Folder deleted successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Folder is not empty
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete folder
      tags:
      - Files
  /folders/batch-stats:
    post:
      consumes:
      - application/json
      description: Get statistics for multiple folders in a single request (max 50
        folders)
      parameters:
      - description: List of folder paths
        in: body
        name: request
        required: true
        schema:
          properties:
            paths:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Folder statistics by path
          schema:
            additionalProperties:
              $ref: '#/definitions/handlers.FolderStats'
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Batch get folder statistics
      tags:
      - Files
  /folders/stats/{path}:
    get:
      consumes:
      - application/json
      description: Get recursive statistics for a folder including file count, folder
        count, and total size
      parameters:
      - description: Folder path
        in: path
        name: path
        required: true
        type: string
      - description: Bypass cache and recompute stats
        in: query
        name: no-cache
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Folder statistics
          schema:
            $ref: '#/definitions/handlers.FolderStats'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Path not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get folder statistics
      tags:
      - Files
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"